Logic update + styling reset
Added in js logic to the ElectionDates vue component. Election dates
and types are now pulled out from the api response and they sit in
their own clear array. I now have automatically updating variables for:
the number of upcoming elections, and the date and type of the next
election.