enhanced logic for next election date + adding address input field
due to the way the API response is structured, I was previously
returning too much text for the next election type (e.g. the special
election type was returned as “Special Election 59 62 and 65
Assembly”.). The logic is now enhanced such that I only return “Special
Election”, without all the stuff after it.